+++ planning/behavior_velocity_planner/autoware_behavior_velocity_no_stopping_area_module/src/scene_no_stopping_area.hpp
@@ -15,6 +15,8 @@
 #ifndef SCENE_NO_STOPPING_AREA_HPP_
 #define SCENE_NO_STOPPING_AREA_HPP_
 
+#include "utils.hpp"
+
 #include <autoware/behavior_velocity_planner_common/scene_module_interface.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/boost_geometry_helper.hpp>
 #include <autoware/behavior_velocity_planner_common/utilization/state_machine.hpp>
@@ -26,24 +28,12 @@
 #include <tier4_planning_msgs/msg/path_with_lane_id.hpp>
 
 #include <memory>
-#include <optional>
-#include <vector>
 
 namespace autoware::behavior_velocity_planner
 {
 class NoStoppingAreaModule : public SceneModuleInterface
 {
 public:
-  struct DebugData
-  {
-    double base_link2front;
-    std::vector<geometry_msgs::msg::Pose> stop_poses;
-    geometry_msgs::msg::Pose first_stop_pose;
-    std::vector<geometry_msgs::msg::Point> stuck_points;
-    geometry_msgs::msg::Polygon stuck_vehicle_detect_area;
-    geometry_msgs::msg::Polygon stop_line_detect_area;
-  };
-
   struct PlannerParam
   {
     /**
@@ -77,8 +67,7 @@ public:
 private:
   const int64_t lane_id_;
 
-  mutable bool pass_judged_ = false;
-  mutable bool is_stoppable_ = true;
+  no_stopping_area::PassJudge pass_judge_;
   StateMachine state_machine_;  //! for state
 
   /**
@@ -91,47 +80,6 @@ private:
     const Polygon2d & poly,
     const autoware_perception_msgs::msg::PredictedObjects::ConstSharedPtr & predicted_obj_arr_ptr);
 
-  /**
-   * @brief Check if there is a stop line in "stop line detect area".
-   * @param path            ego-car lane
-   * @param poly            ego focusing area polygon
-   * @return true if exists
-   */
-  bool check_stop_lines_in_no_stopping_area(
-    const tier4_planning_msgs::msg::PathWithLaneId & path, const Polygon2d & poly);
-
-  /**
-   * @brief Calculate the polygon of the path from the ego-car position to the end of the
-   * no stopping lanelet (+ extra distance).
-   * @param path           ego-car lane
-   * @param ego_pose       ego-car pose
-   * @param margin         margin from the end point of the ego-no stopping area lane
-   * @param extra_dist     extra distance from the end point of the no stopping area lanelet
-   * @return generated polygon
-   */
-  Polygon2d generate_ego_no_stopping_area_lane_polygon(
-    const tier4_planning_msgs::msg::PathWithLaneId & path,
-    const geometry_msgs::msg::Pose & ego_pose, const double margin, const double extra_dist) const;
-
-  /**
-   * @brief Calculate the polygon of the path from the ego-car position to the end of the
-   * no stopping lanelet (+ extra distance).
-   * @param path                  ego-car lane
-   * @param stop_line_margin      stop line margin from the stopping area lane
-   * @return generated stop line
-   */
-  std::optional<universe_utils::LineString2d> get_stop_line_geometry2d(
-    const tier4_planning_msgs::msg::PathWithLaneId & path, const double stop_line_margin) const;
-
-  /**
-   * @brief Calculate if it's possible for ego-vehicle to stop before area consider jerk limit
-   * @param self_pose       ego-car pose
-   * @param line_pose       stop line pose on the lane
-   * @return is stoppable in front of no stopping area
-   */
-  bool is_stoppable(
-    const geometry_msgs::msg::Pose & self_pose, const geometry_msgs::msg::Pose & line_pose) const;
-
   // Key Feature
   const lanelet::autoware::NoStoppingArea & no_stopping_area_reg_elem_;
   std::shared_ptr<const rclcpp::Time> last_obstacle_found_time_;
@@ -140,7 +88,7 @@ private:
   PlannerParam planner_param_;
 
   // Debug
-  DebugData debug_data_;
+  no_stopping_area::DebugData debug_data_;
 };
 }  // namespace autoware::behavior_velocity_planner
